en_GB: Don't match suffixes in en_GB.pl
This fixes the bug, by checking for word boundaries after words to be
translated. This means that suffixes (even the common ones like "s" for
plurals) are never matched. This will require a lot more translations
to be added to the dictionary, but is probably the safest route to take.

If we wanted to cut down on the number of translations added, we could
check for word boundaries, or "s\b", or "es\b" after words to be translated,
to allow for plurals. However, we would have to go through the dictionary
and ensure this was OK for all words, and add exceptions for those words
for which it doesn't make sense. There aren't likely to be any, though.

https://bugzilla.gnome.org/show_bug.cgi?id=574576